Fix queries with limit (#394)
There is an issue with async threads, where we are swallowing the
ClientAbortException when Greenplum closes the connection. This lack of
communication that an IOException has occurred is causing problems with
subsequent requests. Re-throwing the exception allows Spring MVC to
catch it and mark the connection as broken so it stops writing.

The short version of what happens is:
- client drops connection async thread is writing to
- Tomcat cleans this up
- Request/Response/Processor gets put back in pool
- New connections come in
- Processor above gets allocated to one of them
- Original async thread tries writing again, fails and puts the processor into an error state
- Tomcat tries to process the new connection
- Processor is in an error state so the connection is closed before the request is even read

Co-authored-by: Mark Thomas <thomasma@vmware.com>
Co-authored-by: Francisco Guerrero <aguerrero@pivotal.io>